requirements.txt
# Intel Image Classification - Requirements

# Core ML/DL
tensorflow>=2.10.0
numpy>=1.21.0
pandas>=1.3.0

# Image Processing
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Scikit-learn for metrics
scikit-learn>=1.0.0

# Jupyter
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Optional: GPU support (uncomment if using NVIDIA GPU)
# tensorflow-gpu>=2.10.0
